requirements.txt
antiorm==1.2.1
asgiref==3.5.1
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
db==0.1.1
db-sqlite3==0.0.1
Deprecated==1.2.13
Django==4.0.4
django-appconf==1.0.5
django-ckeditor==6.4.2
django-filter==22.1
django-js-asset==2.0.0
django-redis==5.2.0
django-select2==7.10.0
filter==0.0.0.20200724
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
packaging==21.3
Pillow==9.1.1
procfile==0.1.0
pyparsing==3.0.9
python-dateutil==1.5
redis==4.3.4
requests==2.28.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wfastcgi==3.0.0
wrapt==1.14.1
